requirements.txt
PyYAML
numpy==1.18.5
pillow
gym
matplotlib
pytablewriter
opencv-python==4.1.2.30
cloudpickle
pygame
tqdm
imgaug
torch==1.8.0
torchvision==0.9.0
stable-baselines3==1.0
euclid
tensorboard
sklearn
graphviz
imbalanced-learn
statsmodels
mujoco-py==2.0.2.13
pandas
captum==0.4.0
highway-env==1.4
joblib==1.0.1
torchensemble==0.1.7
openpyxl==3.0.9
werkzeug==2.0.3